Refactor data constructor representation
Current implementation assums that a data constructor always has
a single type argument, but the internal representation assumed
a list of arguments. This resulted in lots of warning with pattern
matches, so I changed the internal representation to correspond
to our current assumptions. If we ever decide to change data type
reprsentation this will be simple to undo.